+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Web Session History Viewer
 
+## 2.1.4.0 - 2023-10-22
+* Support for new NINA rolling log files (revised log name matching).
+
+## 2.1.3.0 - 2023-09-24
+* Support for new NINA rolling log files.
+
+## 2.1.2.1 - 2022-12-24
+* Migrated to .NET 7 for NINA 3
+
 ## 1.1.2.1 - 2022-08-07
 * Plugin options are now saved per NINA profile
 * Now supports running another instance of NINA in 'share' mode